modi triangle

+++ src/codility/Triangle.java
@@ -8,7 +8,7 @@ public class Triangle {
         Arrays.sort(A);
 
         for (int i = 0, len = A.length; i < len - 2; i++) {
-            if (A[i] + A[i + 1] > A[i + 2]) {
+            if ((double)A[i] + (double)A[i + 1] > A[i + 2]) {
                 return 1;
             }
         }
